% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{astm_solar_data}
\alias{astm_solar_data}
\title{ASTM G173-03 Reference Solar Spectra Derived from SMARTS v. 2.9.2}
\format{
A data frame with 1701 rows and 2 variables:
\describe{
\item{wavelength}{Wavelength in nanometers (nm)}
\item{w_m2}{Irradiance value (W/m^2/nm)}

}
}
\source{
\url{https://store.astm.org/g0173-03r20.html}
}
\usage{
astm_solar_data
}
\description{
A dataset containing the standard table for reference solar spectral irradiance:
Direct + circumsolar
}
\keyword{datasets}
