% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_color_fractions.R
\name{calc_color_fractions}
\alias{calc_color_fractions}
\title{Color fraction calculation}
\usage{
calc_color_fractions(df, value_col, wavelength_col, exclude_colors = NULL)
}
\arguments{
\item{df}{a dataframe object with spectroradiometer data}

\item{value_col}{the column in the dataframe with the intensity emission data. Must be \strong{numeric}}

\item{wavelength_col}{the column in the dataframe with the wavelength data. Must be \strong{numeric} and wavelength increments
\strong{must} be equal between all measurements.}

\item{exclude_colors}{colors can be manually excluded from color fraction calculation if they are outside the scope
of wavelengths the user is interested in including}
}
\value{
A list item containing the \code{summary} object with of the sum amounts in each color range and their percent contribtions
to the total and \code{parsed_colors} object with the transformed dataframe with negative values removed and with trapezoidal
estimates (if argument = TRUE) and the assigned color for each row of data.
}
\description{
Takes spectroradiometer data and calculates the contributions of ultraviolet-c (uv-c, 100-280 nm),
ultraviolet-b (uv-b, 280-315 nm), ultraviolet-a, (uv-a, 315-400 nm), blue (400-500 nm), green (500-600 nm),
red (600-700 nm), far-red A (fr-a, 700-750 nm), far red B (fr-b, 750-800 nm), and near infrared (nir, 800-2500 nm) using
trapezoidal integration.
}
\examples{
color_summary <- calc_color_fractions(df = astm_solar_data,  value_col = watts_m2, wavelength_col = wavelength,
                                      calculate_trapz_est = FALSE, exclude_colors = c("uv-c","uv-b","nir"))
summary_table <- color_summary$summary
parsed_datafram <- color_summary$parsed_colors

}
