% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{led_spectrum_data}
\alias{led_spectrum_data}
\title{LED Spectrum Data}
\format{
A data frame with 481 rows and 2 variables:
\describe{
\item{wavelength}{Wavelength in nm}
\item{w_m2}{Irradiance value (W/m^2/nm)}
}
}
\usage{
led_spectrum_data
}
\description{
Spectral distribution data for a 9-channel LED used in horticultural lighting.
}
\keyword{datasets}
