% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umol_to_watts.R
\name{umol_to_watts}
\alias{umol_to_watts}
\title{Micromoles to Watts conversion}
\usage{
umol_to_watts(df, value_col, wavelength_col)
}
\arguments{
\item{df}{a dataframe containing spectroradiometer data}

\item{value_col}{the column in the dataframe with the intensity emission data. Must be \strong{numeric}}

\item{wavelength_col}{the column in the dataframe with the wavelength data. Must be \strong{numeric}}
}
\value{
The spectroradiometer reading in Watts/meter^2/nm
}
\description{
Convert Micromols/meter^2/second/nm to Watts/meter^2/nm
}
\examples{
\dontrun{umol_converted <- umol_to_watts(df = led_spectrum_data,
                                         value_col = umol_m2_s,
                                         wavelength_col = wavelength)}
}
