% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watts_to_umol.R
\name{watts_to_umol}
\alias{watts_to_umol}
\title{Watts to micromoles conversion}
\usage{
watts_to_umol(df, value_col, wavelength_col)
}
\arguments{
\item{df}{a dataframe containing spectroradiometer data}

\item{value_col}{the column in the dataframe with the intensity emission data. Must be \strong{numeric}}

\item{wavelength_col}{the column in the dataframe with the wavelength data. Must be \strong{numeric}}
}
\value{
The spectroradiometer reading in Micromols/meter^2/second/nm
}
\description{
Convert Watts/meter^2/nm to Micromols/meter^2/second/nm
}
\examples{
umol_converted <- watts_to_umol(data = astm_solar_data, value_col = w_m2, wavelength_col = wavelength)
}
